requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
numpy>=1.21.0
pyyaml>=6.0
tensorboard>=2.10.0
captum>=0.6.0